+++ armada_command/command_develop.py
@@ -12,9 +12,6 @@ def add_arguments(parser):
                         default=os.path.basename(os.getcwd()),
                         help='Name of the microservice to develop. '
                              'Default is the name of current directory.')
-    parser.add_argument('-P', '--dynamic_ports', action='store_true',
-                        help='Assign dynamic ports, otherwise it will use sticky port in range 4000..4999, based '
-                             'on hash from microservice_name.')
     parser.add_argument('-v', '--volume', default=os.getcwd(),
                         help='Volume mounted to /opt/MICROSERVICE_NAME. Default is current directory path.')
     parser.add_argument('--off', action='store_true',
@@ -25,13 +22,12 @@ def get_armada_develop_env_file_path():
     return '/tmp/armada_develop_env_{}.json'.format(os.getppid())
 
 
-def save_dev_env_vars(microservice_name, dynamic_ports, microservice_volume):
+def save_dev_env_vars(microservice_name, microservice_volume):
     path = get_armada_develop_env_file_path()
     with open(path, 'w') as f:
         env_vars = {
             'ARMADA_DEVELOP': '1',
             'MICROSERVICE_NAME': microservice_name or '',
-            'MICROSERVICE_DYNAMIC_PORTS': '1' if dynamic_ports else '0',
             'MICROSERVICE_VOLUME': microservice_volume or '',
         }
         json.dump(env_vars, f)
@@ -40,7 +36,6 @@ def save_dev_env_vars(microservice_name, dynamic_ports, microservice_volume):
 def command_develop(args):
     off = args.off
     microservice_name = args.microservice_name
-    dynamic_ports = args.dynamic_ports
     volume = args.volume
 
     if off:
@@ -56,4 +51,4 @@ def command_develop(args):
                 'microservice name "{}".'.format(current_dir_name, microservice_name)
               + style.RESET)
 
-    save_dev_env_vars(microservice_name, dynamic_ports, volume)
+    save_dev_env_vars(microservice_name, volume)